Allow PR apps to use it as well without defining an APP_DOMAIN variable
The application domain for the PR apps is built with the
`HEROKU_APP_NAME` env var [1] as the subdomain and "herokuapp.com" as
domain. However, the `HEROKU_APP_NAME` is not available on build time
and we can't use it on the buildpack. So we are sending to the error
tracking tool just "herokuapp.com" as the domain wich the application is
running.

[1] https://devcenter.heroku.com/articles/github-integration-review-apps#injected-environment-variables