Fix the minified url and support new file names pattern
We were sending the wrong URL for the minified js file. This commit
fixes it by using the correct base URL and the file without the full
path.

We also fixed the script to only upload js files.

Additionally, we have changed the code to support the new pattern we
have for the fingerprinted files and source maps file names. It have
changed when we replaced the webpacker by the new Rails 7 recommended
pipeline [1]

[1] https://github.com/vertragswecker/moritz/commit/21771199e4b370389b723da2971a0fb3f5003362